fix: align client contract and add UI coverage (#6)

+++ Sources/JanuarySDK/Foods/Models/FoodSearchResults.swift
@@ -1,9 +1,9 @@
 /// Results from a food-name search.
 public struct FoodSearchResults: Codable, Hashable, Sendable {
-    public var totalCount: Double
+    public var totalCount: Int
     public var items: [FoodSearchItem]
 
-    public init(totalCount: Double, items: [FoodSearchItem]) {
+    public init(totalCount: Int, items: [FoodSearchItem]) {
         self.totalCount = totalCount
         self.items = items
     }
